server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pyjwt[crypto]>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
